Pre-Dropbox commit
I've (Richard) made it so you can choose from the list of customers and instantiate a customer object with all the info. Will do the same for product and user later (next semester). Created all the Java classes for the entities in the project, with Javadoc. Everything is ready to go, except some documentation left in SeniorProject.